requirements.txt
flask
redis
pycurl
Pillow
czipfile
urllib3